requirements-dev.txt
pytest==7.3.1
pytest-mock==3.10.0
requests==2.28.2
pendulum==2.1.2
apache-airflow==2.5.3
apache-airflow-providers-mongo==3.1.1
pymysql==1.0.3